Extract helper to own file
