+++ src/react/hooks/__tests__/useLoadableQuery/customScalars.test.tsx
@@ -1,4 +1,3 @@
-import type { RenderOptions } from "@testing-library/react";
 import { screen } from "@testing-library/react";
 import {
   createRenderStream,
@@ -7,11 +6,9 @@ import {
 } from "@testing-library/react-render-stream";
 import { userEvent } from "@testing-library/user-event";
 import React, { Suspense } from "react";
-import { flushSync } from "react-dom";
-import { ErrorBoundary } from "react-error-boundary";
 import { delay, of } from "rxjs";
 
-import type { DataState, ErrorLike, OperationVariables } from "@apollo/client";
+import type { OperationVariables } from "@apollo/client";
 import {
   ApolloClient,
   ApolloLink,
@@ -36,114 +33,7 @@ import {
 } from "@apollo/client/testing/internal";
 import { invariant } from "@apollo/client/utilities/invariant";
 
-async function renderHook<
-  TData,
-  TVariables extends OperationVariables,
-  TStates extends DataState<TData>["dataState"] = DataState<TData>["dataState"],
-  Props = never,
->(
-  renderHookImpl: (
-    props: Props extends never ? undefined : Props
-  ) => useLoadableQuery.Result<TData, TVariables, TStates>,
-  options: Pick<RenderOptions, "wrapper"> & { initialProps?: Props }
-) {
-  function UseReadQuery({
-    queryRef,
-  }: {
-    queryRef: QueryRef<TData, TVariables, TStates>;
-  }) {
-    useTrackRenders({ name: "useReadQuery" });
-    mergeSnapshot({ result: useReadQuery(queryRef) });
-
-    return null;
-  }
-
-  function SuspenseFallback() {
-    useTrackRenders({ name: "<Suspense />" });
-
-    return null;
-  }
-
-  function ErrorFallback() {
-    useTrackRenders({ name: "<ErrorBoundary />" });
-
-    return null;
-  }
-
-  type RefetchFunction = useLoadableQuery.Handlers<
-    TData,
-    TVariables
-  >["refetch"];
-
-  function App({ props }: { props: Props | undefined }) {
-    useTrackRenders({ name: "useLoadableQuery" });
-    const [loadQuery, queryRef, { refetch }] = renderHookImpl(props as any);
-
-    mergeSnapshot({ loadQuery, refetch });
-
-    return (
-      <Suspense fallback={<SuspenseFallback />}>
-        <ErrorBoundary
-          FallbackComponent={ErrorFallback}
-          onError={(error) => replaceSnapshot({ error })}
-        >
-          {queryRef && <UseReadQuery queryRef={queryRef} />}
-        </ErrorBoundary>
-      </Suspense>
-    );
-  }
-
-  const {
-    render,
-    getCurrentRender,
-    takeRender,
-    mergeSnapshot,
-    replaceSnapshot,
-  } = createRenderStream<
-    | {
-        loadQuery: useLoadableQuery.LoadQueryFunction<TVariables>;
-        refetch: RefetchFunction;
-        result?: useReadQuery.Result<TData, TStates>;
-      }
-    | { error: ErrorLike }
-  >({
-    // These values should always be available, but createRenderStream needs an
-    // initial snapshot when using mergeSnapshot so we provide it with something
-    initialSnapshot: { loadQuery: null as any, refetch: null as any },
-  });
-
-  const utils = await render(<App props={options.initialProps} />, options);
-
-  function rerender(props: Props) {
-    return utils.rerender(<App props={props} />);
-  }
-
-  function getCurrentSnapshot() {
-    const { snapshot } = getCurrentRender();
-    invariant(
-      "loadQuery" in snapshot,
-      "Expected rendered hook instead of error boundary"
-    );
-
-    return snapshot;
-  }
-
-  // React 18 skips committing the suspense fallback when loadQuery/refetch is
-  // triggered as a default-priority update, so the fallback render is missed
-  // and tests may fail. flushSync forces a synchronous commit so the fallback
-  // renders in both React 18 and 19, while returning the underlying result.
-  const loadQuery: useLoadableQuery.LoadQueryFunction<TVariables> = (
-    ...args
-  ) => {
-    return flushSync(() => getCurrentSnapshot().loadQuery(...args));
-  };
-
-  const refetch: RefetchFunction = (...args) => {
-    return flushSync(() => getCurrentSnapshot().refetch(...args));
-  };
-
-  return { takeRender, rerender, getCurrentSnapshot, refetch, loadQuery };
-}
+import { renderUseLoadableQueryHook } from "./testUtils.js";
 
 test("serializes scalar variables used in field arguments", async () => {
   let requestVariables!: OperationVariables;
@@ -478,7 +368,7 @@ test("preserves referential identity when refetching identical serialized scalar
   });
 
   using _disabledAct = disableActEnvironment();
-  const { takeRender, refetch, loadQuery } = await renderHook(
+  const { takeRender, refetch, loadQuery } = await renderUseLoadableQueryHook(
     () => useLoadableQuery(query),
     { wrapper: createClientWrapper(client) }
   );
@@ -609,7 +499,7 @@ test("serializes scalar fields in the error with a `none` error policy", async (
   });
 
   using _disabledAct = disableActEnvironment();
-  const { takeRender, loadQuery } = await renderHook(
+  const { takeRender, loadQuery } = await renderUseLoadableQueryHook(
     () => useLoadableQuery(query, { errorPolicy: "none" }),
     { wrapper: createClientWrapper(client) }
   );
@@ -701,7 +591,7 @@ test("parses scalar fields in the result and serializes them in the error with a
   });
 
   using _disabledAct = disableActEnvironment();
-  const { takeRender, loadQuery } = await renderHook(
+  const { takeRender, loadQuery } = await renderUseLoadableQueryHook(
     () => useLoadableQuery(query, { errorPolicy: "all" }),
     { wrapper: createClientWrapper(client) }
   );
@@ -805,7 +695,7 @@ test("parses custom scalar fields with an `ignore` error policy", async () => {
   });
 
   using _disabledAct = disableActEnvironment();
-  const { takeRender, loadQuery } = await renderHook(
+  const { takeRender, loadQuery } = await renderUseLoadableQueryHook(
     () => useLoadableQuery(query, { errorPolicy: "ignore" }),
     { wrapper: createClientWrapper(client) }
   );
@@ -880,7 +770,7 @@ test("parses custom scalar fields across `@defer` payloads (defer20220824)", asy
   `;
 
   using _disabledAct = disableActEnvironment();
-  const { takeRender, loadQuery } = await renderHook(
+  const { takeRender, loadQuery } = await renderUseLoadableQueryHook(
     () => useLoadableQuery(query),
     { wrapper: createClientWrapper(client) }
   );
@@ -990,7 +880,7 @@ test("parses custom scalar fields across `@defer` payloads (graphql17Alpha9)", a
   `;
 
   using _disabledAct = disableActEnvironment();
-  const { takeRender, loadQuery } = await renderHook(
+  const { takeRender, loadQuery } = await renderUseLoadableQueryHook(
     () => useLoadableQuery(query),
     { wrapper: createClientWrapper(client) }
   );
@@ -1098,7 +988,7 @@ test("parses custom scalar fields across `@stream` payloads (defer20220824)", as
   `;
 
   using _disabledAct = disableActEnvironment();
-  const { takeRender, loadQuery } = await renderHook(
+  const { takeRender, loadQuery } = await renderUseLoadableQueryHook(
     () => useLoadableQuery(query),
     { wrapper: createClientWrapper(client) }
   );
@@ -1225,7 +1115,7 @@ test("parses custom scalar fields across `@stream` payloads (graphql17Alpha9)", 
   `;
 
   using _disabledAct = disableActEnvironment();
-  const { takeRender, loadQuery } = await renderHook(
+  const { takeRender, loadQuery } = await renderUseLoadableQueryHook(
     () => useLoadableQuery(query),
     { wrapper: createClientWrapper(client) }
   );
